[FEATURE] Auto update (#23)
## 1) Description

Added a workflow to automatically update the repositories created from
this template.

## 2) Technical choice
I used the action from
[AndreasAugustin/actions-template-sync](https://github.com/AndreasAugustin/actions-template-sync)
over others alternative. I found this solution was documented enough and
seems to be well-supported. An interesting feature was the ability to
ignore files or folder.

I used a workflow due to their ability to run on schedule and the fact
they are centralized. This centralized behaviour will reduce merge
conflict if we compare with other options like the one mentioned in #21