[BUG]Hot fixes (#30)
## 1) Description

Fixed #5 
Fixed #22 
Resolves #28 
Resolves #29 
Added the `strictBindCallApply` flag to the TS compiler.

## 2) Technical choice

I banned several types as long as the `Function` one. Mostly derivatives
from this type that have no real use cases unless you are working on the
compiler. The `object` type has been banned as well, since this type is
really permissive.

The `any` type as been banned through the usage of the
[`@typescript-eslint/ban-types`](https://typescript-eslint.io/rules/ban-types/)
rule, as it allowed to propose a quick fix. The fix suggested is the
usage of the unknown type, which is almost the same as the `any` but
with type safety. This decision has been made to allow the developers to
use a bit of logic when casting objects to a proper type.