[REFACTOR] ESLint (#32)
## 1) Description
- Resolves #31 

Removed ESLint from the pre commit hooks. It was a bad idea as it
prevent from committing if there is a problem.
Made typing mandatory when assigning to a variable. Note that variables
that hold functions are not concerned by this.
When the types are inferred, it is now possible to let the parameters
non-read-only.