[BUG] Hotfixes (#41)
## 1) Description

- Fixed 34
- Fixed 36
- Fixed 37
- Fixed 38

## 2) Technical choice

Object literal properties do not have to follow a specific set of rules
anymore. Since we enforce the use of interfaces and types everywhere,
the only objects that would not be following our rules would only be the
ones that came from outside our project.
All properties that contain a function now follow the naming convention
of the functions. Hopefully avoiding future issues.