[FEATURE] Better coverage reports (#40)
## 1) Description

Improved how the coverage is displayed in the PR comments. Fixed a bug
that made the coverage not get computed correctly when using locally.

## 2) Technical choice

I chose to use this
[action](https://github.com/marketplace/actions/jest-coverage-comment)
to produce the report. While it doesn't perfectly fit our needs, it does
a good enough jobs and seems to be well maintained.

Other actions have been tested but either did not work at all (one was
yelling at the JSX syntax ...) or lack 90% of the features we wanted to
use.

The issue I fixed is a bug from
[jest](https://github.com/facebook/jest/issues/8562). Since I do not
think this issues will be fixed any time soon, we will have to deal with
this ...