[REFACTOR] tweaked an ESLint rule and a workflow (#46)
## 1) Description

Changed the ESLint rule that prevented the dev from returning an awaited
promise to require it. Awaiting a promise before returning it might seem
useless, but it helps with the stack trace being more precise. This
makes debugging easier.

Jest might return with an exit code of 0 even if some tests failed. This
behavior only exhibits on GH's action system, making it incredibly hard
to find. I added a bit of code to make the workflows fails if a test
fails.

## 2) Technical choice

I used an output from the reporter action. This output is computed from
the output of jest, thus avoiding parsing it ourselves.